Revert "Revert "Devdocs 3880""
